Refactor learning path API hooks out of learning resources
