+++ frontends/main/src/page-components/ManageListDialogs/ManageListDialogs.tsx
@@ -16,9 +16,11 @@ import * as Yup from "yup"
 import { PrivacyLevelEnum, type LearningPathResource, UserList } from "api"
 
 import {
-  useLearningpathCreate,
-  useLearningpathUpdate,
-  useLearningpathDestroy,
+  useLearningPathCreate,
+  useLearningPathUpdate,
+  useLearningPathDestroy,
+} from "api/hooks/learningPaths"
+import {
   useLearningResourceTopics,
   useUserListCreate,
   useUserListUpdate,
@@ -82,8 +84,8 @@ const UpsertLearningPathDialog = NiceModal.create(
     const topicsQuery = useLearningResourceTopics(undefined, {
       enabled: modal.visible,
     })
-    const createList = useLearningpathCreate()
-    const updateList = useLearningpathUpdate()
+    const createList = useLearningPathCreate()
+    const updateList = useLearningPathUpdate()
     const mutation = resource?.id ? updateList : createList
     const handleSubmit: FormikConfig<
       LearningPathResource | LearningPathFormValues
@@ -301,7 +303,7 @@ const DeleteLearningPathDialog = NiceModal.create(
   ({ resource }: DeleteLearningPathDialogProps) => {
     const modal = NiceModal.useModal()
     const hideModal = modal.hide
-    const destroyList = useLearningpathDestroy()
+    const destroyList = useLearningPathDestroy()
 
     const handleConfirm = useCallback(async () => {
       await destroyList.mutateAsync({
